+++ gulp/compendiumSchemas.ts
@@ -1,5 +1,6 @@
 import { promises as fs } from "fs";
 import { COMP_AMMO_JSON_SCHEMA } from "../src/main/typescript/data/item/ammo/source.js";
+import { COMP_APPAREL_JSON_SCHEMA } from "../src/main/typescript/data/item/apparel/source.js";
 import { COMP_WEAPON_JSON_SCHEMA } from "../src/main/typescript/data/item/weapon/source.js";
 
 // The paths here are relative to the project root
@@ -11,6 +12,11 @@ const schemaConfigs: SchemaConfig[] = [
     fileName: "ammo",
     outputBasePath: itemOutputBasePath,
     schema: COMP_AMMO_JSON_SCHEMA
+  },
+  {
+    fileName: "apparel",
+    outputBasePath: itemOutputBasePath,
+    schema: COMP_APPAREL_JSON_SCHEMA
   },
   {
     fileName: "weapon",