+++ src/main/typescript/config.ts
@@ -7,6 +7,7 @@ import WvItemSheet from "./applications/item/wvItemSheet.js";
 import { CONSTANTS, TYPES } from "./constants.js";
 import { CHARACTER_JSON_SCHEMA } from "./data/actor/character/source.js";
 import { AMMO_SOURCE_JSON_SCHEMA } from "./data/item/ammo/source.js";
+import { APPAREL_SOURCE_JSON_SCHEMA } from "./data/item/apparel/source.js";
 import { BASE_ITEM_JSON_SCHEMA } from "./data/item/baseItem.js";
 import { STACK_BASE_ITEM_JSON_SCHEMA } from "./data/item/stackableBaseItem.js";
 import { WEAPON_SOURCE_JSON_SCHEMA } from "./data/item/weapon/source.js";
@@ -46,6 +47,7 @@ export function configureFoundryOnInit(): void {
       },
       item: {
         ammo: ajv.compile(AMMO_SOURCE_JSON_SCHEMA),
+        apparel: ajv.compile(APPAREL_SOURCE_JSON_SCHEMA),
         effect: ajv.compile(BASE_ITEM_JSON_SCHEMA),
         misc: ajv.compile(STACK_BASE_ITEM_JSON_SCHEMA),
         weapon: ajv.compile(WEAPON_SOURCE_JSON_SCHEMA)