+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - mocha and chai testing framework
 - a dedicated caps field for actors
   ([#193](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/193))
+- basic support for apparel items
+  ([#140](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/140))
 
 ### Changed